pyproject.toml
[project]
name = "rag-project"
version = "0.2.0"
description = "Production-ready RAG pipeline with FastAPI, LangChain, and PGVector"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # LangChain ecosystem
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langchain-postgres>=0.0.13",
    # Document parsing
    "pypdf>=6.9.1",
    "docx2txt>=0.8",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.18",
    "slowapi>=0.1.9",
    # Config & logging
    "pydantic-settings>=2.0",
    "python-dotenv>=1.2.2",
    "structlog>=24.0",
    # Database
    "psycopg[binary]>=3.2",
    "sqlalchemy>=2.0",
    # Caching
    "cachetools>=5.5",
    # Observability
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48",
    "prometheus-fastapi-instrumentator>=7.0",
]

[project.scripts]
rag-cli = "rag.cli:main"
rag-server = "rag.api.app:run_server"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
